2023-09-12-sj-003 심리검사시작페이지 수정중 검사시작 페이지 푸터 수정해야 함
