+++ src/main/resources/static/css/match/therapyList.css
@@ -12,6 +12,26 @@ body{
 .search{
     flex: 1;
 }
+#scroll-btn {
+    cursor: pointer;
+    opacity: 0;
+    width: 40px;
+    height: 40px;
+    color: #fff;
+    background-color: #C09A81;
+    position: fixed;
+    top: 10%;
+    right: 7%;
+    border: 2px solid #fff;
+    border-radius: 50%;
+    font: bold 20px monospace;
+    transition: opacity 0.5s, transform 0.5s;
+}
+
+#scroll-btn.show {
+    opacity: 1;
+    transition: opacity 1s, transform 1s;
+}
 
 /* 검색 */
 .search-bar * {
