+++ src/main/resources/static/css/match/therapyRecommend.css
@@ -4,6 +4,15 @@ body{
     background: #FFF6E2;
     text-align: center;
     margin:0;
+    height: 100%;
+}
+.grass{
+    position: fixed;
+    left: 0;
+    bottom: 0;
+    width: 100%;
+    height: 200px;
+    margin-top: -300px;
 }
 .parent{
     font-family: 'Noto Sans KR';
@@ -25,7 +34,7 @@ body{
     display: inline-block;
     height: 130px;
     width: 130px;
-    margin-bottom: 50px;
+    margin-bottom: 30px;
     z-index: 1;
 }
 
@@ -82,7 +91,7 @@ body{
     width: 100%;
     padding: 0;
     margin: 0;
-    bottom: -100px;
+/*    bottom: -100px;*/
     left: 0;
     transition: all 0.5s ease 0s;
 }
