+++ src/main/resources/static/css/psychological/lastTest.css
@@ -16,13 +16,14 @@ textarea {
     width: 100%;
     height: 200px;
     margin-top: -300px;
+    z-index: -1;
 }
 .box{
     margin: 40px auto;
     text-align: center;
     box-sizing: border-box;
     width: 500px;
-    height: 500px;
+    height: 550px;
     background: #FFFFFF;
     border: 1px solid rgba(192, 154, 129, 0.8);
     border-radius: 10px;
