+++ src/main/resources/static/css/psychological/startTest.css
@@ -13,11 +13,11 @@ span{
     font-size: 23px;
     font-weight: bold;
     color: white;
-    margin: 15%;
+    margin: 7%;
 }
 .countbutton{
-    width: 100px;
-    height: 50px;
+    width: 110px;
+    height: 60px;
     color: #C09A81;
     background: #FFFFFF;
     border-radius: 30px;
@@ -30,10 +30,11 @@ span{
     width: 100%;
     height: 200px;
     margin-top: -300px;
+    z-index: -1;
 }
 .picture{
-    width: auto;
-    height: auto;
+    width: 25%;
+    height: 25%;
 }
 .whitrtext{ color: white;}
