+++ src/main/resources/static/css/terms/termsOfService.css
@@ -10,7 +10,6 @@ body{
     display: flex;
     width: 1000px;
     height: 700px;
-    overflow:scroll;
     text-align: center;
     word-break:normal;
     word-break:keep-all;
@@ -18,4 +17,13 @@ body{
 .titleLink {
     text-decoration: none;
     color: inherit;
+}
+textarea{
+    color: #C09A81;
+    font-weight: bold;
+    background-color: #FFF6E2;
+    width: 100%;
+    height: 100%;
+    border: none;
+    resize: none;
 }
