+++ src/main/resources/templates/Components/PreLoginHeader.html
@@ -81,7 +81,7 @@
 
     <ul id="menu">
         &emsp;&emsp;<li><a href="/introduction/introduction">소개</a></li>
-        &emsp;&emsp;<li><a href="/psychological/therapylist">상담사</a></li>
+        &emsp;&emsp;<li><a href="/match/therapyList">상담사</a></li>
         &emsp;&emsp;<li><a class="titleLink" href="/review/reviewmain">후기</a></li>
         &emsp;&emsp;<li>FAQ</li>
         &emsp;&emsp;<li><a href="/regist/login">로그인</a></li>
