+++ src/main/resources/templates/match/introduceTherapy.html
@@ -3,7 +3,8 @@
 <head>
     <meta charset="UTF-8">
     <title>Introduce</title>
-    <link rel="stylesheet" type="text/css" href="../css/psychological/introduceTherapy.css">
+    <link rel="stylesheet" type="text/css" href="../css/match/introduceTherapy.css">
+    <link href="https://fonts.googleapis.com/css2?family=Noto+Sans+KR:wght@100;300;400;500;700;900&display=swap" rel="stylesheet">
 </head>
 <body>
     <header>
