+++ src/main/resources/templates/match/therapyRecommend.html
@@ -3,7 +3,8 @@
 <head>
     <meta charset="UTF-8">
     <title>list</title>
-    <link rel="stylesheet" type="text/css" href="../css/psychological/therapyRecommend.css">
+    <link rel="stylesheet" type="text/css" href="../css/match/therapyRecommend.css">
+    <link href="https://fonts.googleapis.com/css2?family=Noto+Sans+KR:wght@100;300;400;500;700;900&display=swap" rel="stylesheet">
 </head>
 <body>
 <th:block th:include="Components/PostLoginHeader.html"></th:block>
@@ -19,7 +20,7 @@
                 <h4>전문 분야</h4>
             </div>
             <ul class="social">
-                <li><a href="/psychological/introduceTherapy" class="profilebutton" aria-hidden="true">프로필 보기</a></li>
+                <li><a href="/match/introduceTherapy" class="profilebutton" aria-hidden="true">프로필 보기</a></li>
             </ul>
         </div>
         <div class="our-team">
@@ -32,7 +33,7 @@
                 <h4>전문 분야</h4>
             </div>
             <ul class="social">
-                <li><a href="/psychological/introduceTherapy" class="profilebutton" aria-hidden="true">프로필 보기</a></li>
+                <li><a href="/match/introduceTherapy" class="profilebutton" aria-hidden="true">프로필 보기</a></li>
             </ul>
         </div>
         <div class="our-team">
@@ -45,7 +46,7 @@
                 <h4>전문 분야</h4>
             </div>
             <ul class="social">
-                <li><a href="/psychological/introduceTherapy" class="profilebutton" aria-hidden="true">프로필 보기</a></li>
+                <li><a href="/match/introduceTherapy" class="profilebutton" aria-hidden="true">프로필 보기</a></li>
             </ul>
         </div>
         <div class="our-team">
@@ -58,7 +59,7 @@
                 <h4>전문 분야</h4>
             </div>
             <ul class="social">
-                <li><a href="/psychological/introduceTherapy" class="profilebutton" aria-hidden="true">프로필 보기</a></li>
+                <li><a href="/match/introduceTherapy" class="profilebutton" aria-hidden="true">프로필 보기</a></li>
             </ul>
         </div>
         <div class="our-team">
@@ -71,13 +72,13 @@
                 <h4>전문 분야</h4>
             </div>
             <ul class="social">
-                <li><a href="/psychological/introduceTherapy" class="profilebutton" aria-hidden="true">프로필 보기</a></li>
+                <li><a href="/match/introduceTherapy" class="profilebutton" aria-hidden="true">프로필 보기</a></li>
             </ul>
         </div>
     </div>
 </div>
 <div>
-    <a style="color: #C08181; text-align: center; font-weight: bold;" href="/psychological/therapylist">또 다른 훌륭한 상담사들 보기</a>
+    <a style="color: #C08181; text-align: center; font-weight: bold;" href="/match/therapyList">또 다른 훌륭한 상담사들 보기</a>
 </div>
 <footer>
     <th:block th:include="Components/Footer.html" th:with="textColor='#C09A81'"></th:block>
