+++ src/main/resources/templates/mypage/user/profile1.html
@@ -32,7 +32,7 @@
         </tr>
     </table>
     <br><br>
-    <input type="button" value="상담사 보기" onclick="location.href='../psychological/therapylist'" style="width: 200px; height: 33px; border-radius: 10px; background-color: #C09A81; color: #FFFFFF; border-color: #C09A81; border-width: 1px; border: solid; cursor: pointer;"><br><br>
+    <input type="button" value="상담사 보기" onclick="location.href='../match/therapyList'" style="width: 200px; height: 33px; border-radius: 10px; background-color: #C09A81; color: #FFFFFF; border-color: #C09A81; border-width: 1px; border: solid; cursor: pointer;"><br><br>
     <input type="button" value="검사 하러 가기" onclick="location.href='../psychological/test'" style="width: 200px; height: 33px; border-radius: 10px; background-color: #FFFFFF; color: #C09A81; border-color: #C09A81; border-width: 1px; border: solid; cursor: pointer;">
 </div>
 </script>
