+++ src/main/resources/templates/psychological/firstTest.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <title>first</title>
     <link rel="stylesheet" type="text/css" href="../css/psychological/icon.css">
-
+    <link href="https://fonts.googleapis.com/css2?family=Noto+Sans+KR:wght@100;300;400;500;700;900&display=swap" rel="stylesheet">
     <link rel="stylesheet" type="text/css" href="/css/psychological/firstTest.css">
 </head>
 <body>
