+++ src/main/resources/templates/psychological/lastTest.html
@@ -4,6 +4,7 @@
     <meta charset="UTF-8">
     <title>last</title>
     <link rel="stylesheet" type="text/css" href="../css/psychological/lastTest.css">
+    <link href="https://fonts.googleapis.com/css2?family=Noto+Sans+KR:wght@100;300;400;500;700;900&display=swap" rel="stylesheet">
 </head>
 <body>
 <header>
@@ -13,7 +14,7 @@
 <div class="box"><br>
     해당사항이 있으면 체크해주세요.<br><br>
     <h3>희망 상담사 성별</h3>
-    <form action="/psychological/match" method="get"><br>
+    <form action="/psychological/loding" method="get"><br>
         <input type="checkbox" name="gender" value="남">남자<br>
         <input type="checkbox" name="gender" value="여">여자<br><br>
     <h3>최근고민분야</h3><br>
@@ -23,7 +24,7 @@
     <input type="checkbox" name="alone" value="소외감">소외감
     <input type="checkbox" name="etc" value="기타">기타 <br><br>
         <textarea placeholder="기타내용작성 50장 이내" style="border-radius: 10px; border-color: #C09A81"></textarea> <br>
-        <input class="roundbutton" type="submit" value="제출하기">
+        <input class="roundbutton" type="submit" value="제출하기"><br>
     </form>
 </div>
 <footer>
