+++ src/main/resources/templates/psychological/lodingPage.html
@@ -3,6 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <title>matching</title>
+    <link href="https://fonts.googleapis.com/css2?family=Noto+Sans+KR:wght@100;300;400;500;700;900&display=swap" rel="stylesheet">
     <style>
         body{
             justify-content: center; /* 가로 중앙 정렬 */
