+++ src/main/resources/templates/psychological/resultTest.html
@@ -4,6 +4,7 @@
     <meta charset="UTF-8">
     <title>resultTest</title>
     <link rel="stylesheet" type="text/css" href="../css/psychological/resultTest.css">
+    <link href="https://fonts.googleapis.com/css2?family=Noto+Sans+KR:wght@100;300;400;500;700;900&display=swap" rel="stylesheet">
 </head>
 <body style="text-align: center">
     <header>
@@ -11,6 +12,7 @@
         <h1>메타님의 심리상태</h1><th:block th:include="Components/manProfile"></th:block>
         <br>
     </header>
+
     <div class="wrap">
         <div class="wrap img">
             <img class="grass" src="/images/common/Background Shape.png" alt="grass">
@@ -33,7 +35,7 @@
                     내점수 50<br>
                     평균<br>
                 </div>
-            </div>
+            </div><br>
             <div class="explanation">'신경증'은 당신이 얼마나 민감한지 보여주는 성격 요소입니다.
                 '신경증' 점수가 높은 사람일수록 주변 변화에 예민하게 반응해서
                 감정 기복이 심하고 스트레스를 많이 받지만,
@@ -100,7 +102,7 @@
                 </div>
             </div>
             <br>
-            <button class="matchingbutton" onclick="location.href='/psychological/therapyRecommend'"> 매칭 결과 보러가기</button>
+            <button class="matchingbutton" onclick="location.href='/match/therapyRecommend'"> 매칭 결과 보러가기</button>
             <br><br><br>
                 <footer>
                     <th:block th:include="Components/Footer.html" th:with="textColor='#C09A81'"></th:block>
@@ -109,26 +111,3 @@
     </div>
 </body>
 </html>
-
-<!--
- 같이가치 결과
-<div class="graph_compare"><div class="shape_compare"></div>
-    <div class="info_compare my_compare" style="left: 50%;">
-        <div class="indicate_point">
-            <div class="name_info">
-                <span class="txt_compare">내점수&nbsp;
-                    <em class="tit_emph">50</em>
-                </span>
-            </div>
-            <div class="line_info">
-            </div>
-            <div class="dot_info"></div>
-        </div>
-    </div>
-    <div class="info_compare average_compare" style="left: 55.14%;">
-        <div class="indicate_point"><div class="line_info"></div><div class="name_info">
-            <span class="txt_compare">평균</span>
-        </div>
-        </div>
-    </div>
-</div>-->
