+++ src/main/resources/templates/regist/therapistRegistSuccess.html
@@ -23,7 +23,7 @@
 <h1>상담사로 가입이 완료되었습니다!</h1>
 <p>첨부 자료 검토 후<br>활동 승인이 이루어지오니<br>잠시만 기다려주세요</p>
 <div class="buttons">
-<button onclick="location.href='postLoginMain'" class="whiteButton">구경하러 가기</button><button onclick="location.href='사용자 프로필로 가야함'" class="whiteButton">내 프로필 수정하기</button>
+<button onclick="location.href='/'" class="whiteButton">구경하러 가기</button><button onclick="location.href='/therapist/profile'" class="whiteButton">내 프로필 수정하기</button>
 </div>
 </div>
 <th:block th:include="/Components/beigeBackgroundBrownGrass.html"></th:block>
