+++ src/main/resources/templates/terms/termsOfService.html
@@ -4,6 +4,7 @@
     <meta charset="UTF-8">
     <title>termsOfService</title>
     <link rel="stylesheet" type="text/css" href="/css/terms/termsOfService.css">
+    <link href="https://fonts.googleapis.com/css2?family=Noto+Sans+KR:wght@100;300;400;500;700;900&display=swap" rel="stylesheet">
 </head>
 <body>
 <th:block th:include="Components/PostLoginHeader.html" th:with="textColor='#C09A81'"></th:block>
@@ -14,7 +15,7 @@
 </div>
 <h3>이용약관</h3>
 <div class="whitebox">
-    <div>
+    <textarea readonly>
         (주)메타버츄얼 (이하 '회사'라 함)는 개인정보보호법, 통신비밀보호법, 전기통신사업법, 정보통신망 이용촉진 및 정보보호 등에 관한 법률 등 정보통신서비스제공자가 준수하여야 할 관련 법령상의 개인정보보호 규정을 준수하며, 관련 법령에 의거해 개인정보처리방침을 정하여 이용자 권익 보호에 최선을 다하고 있습니다. 회사의 개인정보처리방침은 다음과 같습니다.
 
 
@@ -308,7 +309,7 @@
         우편을 이용하는 경우 내용증명으로 발송
         e-mail 에 의한 자료제공시 업무용 컴퓨터에서 사용하는 문서작성 도구의 자체 암호화 기능 등을 통하여 개인정보 파일을 암호화하여 제공
         자료제공시 반드시 내부결제를 득한 후 제공
-        문서 출력물에 의해 직접 전달하는 경우 요청기관의 담당자 또는 직상급자에게 직접전달(수령자 인적사항 등을 반드시 기록하여 관리)</div>
+        문서 출력물에 의해 직접 전달하는 경우 요청기관의 담당자 또는 직상급자에게 직접전달(수령자 인적사항 등을 반드시 기록하여 관리)</textarea>
 </div>
 <footer>
     <th:block th:include="Components/Footer.html" th:with="textColor='#C09A81'"></th:block>
